+++ core/apis_test.go
@@ -22,8 +22,8 @@ import (
 	genmocks "core-building-block/core/mocks"
 	"core-building-block/core/model"
 
-	"github.com/rokwire/core-auth-library-go/v3/tokenauth"
-	"github.com/rokwire/logging-library-go/v2/logs"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/tokenauth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logs"
 	"github.com/stretchr/testify/mock"
 	"gotest.tools/assert"
 )