+++ driver/web/apis_enc.go
@@ -18,8 +18,8 @@ import (
 	"core-building-block/core"
 	"net/http"
 
-	"github.com/rokwire/core-auth-library-go/v3/tokenauth"
-	"github.com/rokwire/logging-library-go/v2/logs"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/tokenauth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logs"
 )
 
 // EncApisHandler handles the APIs implementation used by the Encryption BB