+++ driver/web/conversions_application.go
@@ -19,8 +19,8 @@ import (
 	Def "core-building-block/driver/web/docs/gen"
 	"core-building-block/utils"
 
-	"github.com/rokwire/logging-library-go/v2/errors"
-	"github.com/rokwire/logging-library-go/v2/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/errors"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
 )
 
 // PartialApp