+++ main.go
@@ -28,14 +28,14 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/rokwire/core-auth-library-go/v3/authservice"
-	"github.com/rokwire/core-auth-library-go/v3/authutils"
+	rokwireAuth "github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/rokwireutils"
 
-	"github.com/rokwire/core-auth-library-go/v3/envloader"
-	"github.com/rokwire/core-auth-library-go/v3/keys"
-	"github.com/rokwire/logging-library-go/v2/errors"
-	"github.com/rokwire/logging-library-go/v2/logs"
-	"github.com/rokwire/logging-library-go/v2/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/keys"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/envloader"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/errors"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logs"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
 )
 
 var (
@@ -184,7 +184,7 @@ func main() {
 	profileBBApiKey := envLoader.GetAndLogEnvVar("ROKWIRE_CORE_PROFILE_BB_API_KEY", false, true)
 	profileBBAdapter := profilebb.NewProfileBBAdapter(migrate, profileBBHost, profileBBApiKey)
 
-	authService := &authservice.AuthService{
+	authService := &rokwireAuth.Service{
 		ServiceID:   serviceID,
 		ServiceHost: host,
 		FirstParty:  true,
@@ -197,7 +197,7 @@ func main() {
 	}
 
 	serviceAccountLoader := auth.NewLocalServiceAccountLoader(*authImpl)
-	serviceAccountManager, err := authservice.NewServiceAccountManager(authService, serviceAccountLoader)
+	serviceAccountManager, err := rokwireAuth.NewServiceAccountManager(authService, serviceAccountLoader)
 	if err != nil {
 		logger.Fatalf("Error initializing service account manager: %v", err)
 	}
@@ -221,7 +221,7 @@ func main() {
 	var envData *model.EnvConfigData
 	var corsAllowedHeaders []string
 	var corsAllowedOrigins []string
-	config, err := storageAdapter.FindConfig(model.ConfigTypeEnv, authutils.AllApps, authutils.AllOrgs)
+	config, err := storageAdapter.FindConfig(model.ConfigTypeEnv, rokwireutils.AllApps, rokwireutils.AllOrgs)
 	if err != nil {
 		logger.Fatal(errors.WrapErrorAction(logutils.ActionFind, model.TypeConfig, nil, err).Error())
 	}